server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
python-dotenv
google-generativeai
